Updated starter-snakes.md with rust starter snake (#41)
The rust starter snake provides a way to unit test the snake and allows users only to think about the logic by doing all the serializing and deserializing for the user. The unit tests make sure that the snake returns HTTP OK and a valid response.

Rust starter snake only requires Rust + Cargo and tests can be run with `cargo test`.